+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,14 @@
 
 ## Next Release
 
+- fix browser node store issue
+
+## 0.0.28
+
+- use receiver payment name instead of transfer definition
+- add protocol-level validation
+- add `create` protocol integration tests
+
 ## 0.0.26
 
 - fix cross-chain transfer bug